Fix a difference in definition of open_fd_count

+++ src/FSAL/FSAL_GPFS/fsal_convert.c
@@ -78,12 +78,12 @@ int posix2fsal_error(int posix_errorcode)
 		/* all shown as IO errors */
 		if (getrlimit(RLIMIT_NOFILE, &rlim) != 0) {
 			LogInfo(COMPONENT_FSAL,
-				"Mapping %d to ERR_FSAL_IO, open_fd_count=%d getrlimit failed",
-				posix_errorcode, open_fd_count);
+				"Mapping %d to ERR_FSAL_IO, open_fd_count=%ld getrlimit failed",
+				posix_errorcode, (long) open_fd_count);
 		} else {
 			LogInfo(COMPONENT_FSAL,
-				"Mapping %d to ERR_FSAL_IO, open_fd_count=%d rlim_cur=%ld rlim_max=%ld",
-				+posix_errorcode, open_fd_count, rlim.rlim_cur,
+				"Mapping %d to ERR_FSAL_IO, open_fd_count=%ld rlim_cur=%ld rlim_max=%ld",
+				+posix_errorcode, (long) open_fd_count, rlim.rlim_cur,
 				rlim.rlim_max);
 		}
 		return ERR_FSAL_IO;
