+++ src/FSAL/FSAL_GPFS/fsal_convert.h
@@ -74,6 +74,6 @@ static inline struct timespec posix2fsal_time(time_t tsec, time_t nsec)
 #define high32m(a) ((unsigned int)(a >> 32))
 #define low32m(a) ((unsigned int)a)
 
-extern uint32_t open_fd_count;
+extern size_t open_fd_count;
 
 #endif
